requirements.txt
django
django-money
djangorestframework
djangorestframework-jsonapi
djangorestframework-jwt
djoser
mysqlclient
django-filter